refactor(guest): encapsulate resolve_user inside container module
Move check_rootfs_ownership into container/mod.rs so resolve_user
never crosses the module boundary. service/container.rs now imports
a single check_rootfs_ownership instead of resolve_user + a local
helper; spec stays fully private.
