+++ src/guest/src/container/mod.rs
@@ -89,6 +89,34 @@ pub use lifecycle::Container;
 #[cfg(target_os = "linux")]
 pub(crate) use mount_override::validate_mount_override;
 #[cfg(target_os = "linux")]
-pub(crate) use spec::resolve_user;
-#[cfg(target_os = "linux")]
 pub use spec::{ContainerDevices, MountOverride, UserMount};
+
+/// Verify that a disk-strategy rootfs has the expected exec-user ownership.
+///
+/// Resolves the exec user from the mounted rootfs's `/etc/passwd`, then
+/// delegates to [`crate::storage::perms::verify_and_repair_ownership`].
+/// Failures are logged as warnings — they do not abort container init.
+#[cfg(target_os = "linux")]
+pub(crate) fn check_rootfs_ownership(rootfs: &std::path::Path, user: &str) {
+    let rootfs_str = rootfs.to_string_lossy();
+    match spec::resolve_user(&rootfs_str, user) {
+        Ok((exec_uid, exec_gid)) => {
+            if let Err(e) =
+                crate::storage::perms::verify_and_repair_ownership(rootfs, exec_uid, exec_gid)
+            {
+                tracing::warn!(
+                    "Ownership verification failed for rootfs {}: {}",
+                    rootfs.display(),
+                    e
+                );
+            }
+        }
+        Err(e) => {
+            tracing::warn!(
+                "Could not resolve exec user '{}' for ownership check: {}",
+                user,
+                e
+            );
+        }
+    }
+}
